avance, solo falta arreglado dos bugs: no se muestra todo en tracker, y no se muestran notas en monitoreo
